Subo actualizacion del sistema

+++ pacientes_2023/views.py
@@ -4,6 +4,7 @@ from .models import Paciente
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth import logout
 from django.contrib import messages
+from django.http import JsonResponse
 
 
 
@@ -26,7 +27,7 @@ def agregar_paciente(request):
     if request.method == 'POST':
         form = PacienteForm(request.POST)
         form.save()
-        messages.success(request, 'Paciente Agregado exitosamente.')
+        messages.success(request, 'Paciente Agregado')
         return redirect('index')
         
     else:
@@ -34,18 +35,21 @@ def agregar_paciente(request):
         
     return render(request, 'agregar_paciente.html', {'form': form})
 
+
 @login_required
 def index(request):
     pacientes = Paciente.objects.all()
     servicios = Paciente.objects.values_list('servicio', flat=True).distinct()
-    return render(request, 'index.html', {'pacientes': pacientes, 'servicios': servicios})
+    pertenece_admin = request.user.groups.filter(name='admins').exists()
+    return render(request, 'index.html', {'pacientes': pacientes, 'servicios': servicios,'pertenece_admin': pertenece_admin,})
 
 @login_required
 def borrar_paciente(request, paciente_id):
     paciente = get_object_or_404(Paciente, id=paciente_id)
-    paciente.delete()
-    messages.success(request, 'Paciente Eliminado exitosamente.')
-    return redirect('index')
+    if request.method == 'POST':
+        paciente.delete()
+        messages.success(request, 'Paciente Eliminado')
+    return redirect('index')   
 
 
 @login_required
@@ -56,7 +60,7 @@ def editar_paciente(request, paciente_id):
         form = PacienteForm(request.POST, instance=paciente)
         if form.is_valid():
             form.save()
-            messages.success(request, 'Paciente Editado exitosamente.')
+            messages.success(request, 'Paciente Editado')
             return redirect('index')
     else:
         form = PacienteForm(instance=paciente)
